requirements.txt
selenium>=4.9.0
beautifulsoup4>=4.11.0
python-dotenv>=1.0.0
google-auth>=2.16.0
google-auth-oauthlib>=1.0.0
google-auth-httplib2>=0.1.0
google-api-python-client>=2.80.0
pytz>=2023.3 